- Executor functions delegate calls to their `WithContext` pair. - These new `WithContext` functions check the context and pass it to the client instead of just using the root context.
